requirements.txt
plotly==4.14.3
streamlit==0.74.1
pandas==1.2.0
openpyxl==3.0.6
Pillow==8.1.0

#### pip install  plotly streamlit pandas openpyxl Pillow
 